Move FuzzFuzzyFind() in package fuzzy_test
